Refactor type imports to use client-schema for improved type safety across components
