+++ amplify/client-schema.ts
@@ -1,0 +1,1 @@
+export type { Schema } from './shared/data-schema.generated';
