+++ src/OrganizationSelector.tsx
@@ -1,6 +1,6 @@
 import { useContext, useEffect, useState } from 'react';
 import { GlobalContext, UserContext } from './Context';
-import { Schema } from '../amplify/data/resource';
+import { Schema } from '../amplify/client-schema';
 import Dropdown from 'react-bootstrap/Dropdown';
 import DropdownButton from 'react-bootstrap/DropdownButton';
 import ButtonGroup from 'react-bootstrap/ButtonGroup';
