+++ src/schemaTypes.tsx
@@ -1,4 +1,4 @@
-import type { Schema } from '../amplify/data/resource';
+import type { Schema } from '../amplify/client-schema';
 
 export type ProjectType = Schema["Project"]["type"];
 export type CategoryType = Schema["Category"]["type"];
